+++ src/Modules/Admin/Controllers/HotspotAssetController.php
@@ -39,7 +39,7 @@ class HotspotAssetController extends Controller
                 'rules' => [
                     [
                         'allow' => true,
-                        'actions' => ['create', 'delete', 'duplicate', 'index', 'order', 'update'],
+                        'actions' => ['create', 'delete', 'duplicate', 'index', 'order', 'status', 'update'],
                         'roles' => [Entry::AUTH_ENTRY],
                     ],
                 ],
@@ -72,6 +72,11 @@ class HotspotAssetController extends Controller
     public function actionDelete(int $id): Response|string
     {
         return $this->deleteAsset($this->findHotspotAsset($id));
+    }
+
+    public function actionStatus(int $id): Response
+    {
+        return $this->updateStatus($this->findHotspotAsset($id));
     }
 
     public function actionDuplicate(int $id): Response|string